requirements.txt
langchain
langchain-huggingface
langchain-pinecone
langchain-groq
langgraph
edgartools